+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/certificationId/Validator20142015.java
@@ -10,135 +10,16 @@ import gov.healthit.chpl.certificationId.Validator;
 
 public class Validator20142015 extends Validator {
 
-	protected static final List<String> INPATIENT_CQMS = new ArrayList<String> (Arrays.asList(
-		"CMS9",
-		"CMS26",
-		"CMS30",
-		"CMS31",
-		"CMS32",
-		"CMS53",
-		"CMS55",
-		"CMS60",
-		"CMS71",
-		"CMS72",
-		"CMS73",
-		"CMS91",
-		"CMS100",
-		"CMS102",
-		"CMS104",
-		"CMS105",
-		"CMS107",
-		"CMS108",
-		"CMS109",
-		"CMS110",
-		"CMS111",
-		"CMS113",
-		"CMS114",
-		"CMS171",
-		"CMS172",
-		"CMS178",
-		"CMS185",
-		"CMS188",
-		"CMS190"
-	));
-
-	protected static final List<String> AMBULATORY_CQMS = new ArrayList<String> (Arrays.asList(
-// Core		"CMS2",
-		"CMS22",
-// Core		"CMS50",
-		"CMS52",
-		"CMS56",
-		"CMS61",
-		"CMS62",
-		"CMS64",
-		"CMS65",
-		"CMS66",
-// Core		"CMS68",
-// Core		"CMS69",
-		"CMS74",
-// Core		"CMS75",
-		"CMS77",
-		"CMS82",
-// Core		"CMS90",
-// Core		"CMS117",
-		"CMS122",
-		"CMS123",
-		"CMS124",
-		"CMS125",
-// Core		"CMS126",
-		"CMS127",
-		"CMS128",
-		"CMS129",
-		"CMS130",
-		"CMS131",
-		"CMS132",
-		"CMS133",
-		"CMS134",
-		"CMS135",
-// Core		"CMS136",
-		"CMS137",
-// Core		"CMS138",
-		"CMS139",
-		"CMS140",
-		"CMS141",
-		"CMS142",
-		"CMS143",
-		"CMS144",
-		"CMS145",
-// Core		"CMS146",
-		"CMS147",
-		"CMS148",
-		"CMS149",
-// Core		"CMS153",
-// Core		"CMS154",
-// Core		"CMS155",
-// Core		"CMS156",
-		"CMS157",
-		"CMS158",
-		"CMS159",
-		"CMS160",
-		"CMS161",
-		"CMS163",
-		"CMS164",
-// Core		"CMS165",
-// Core		"CMS166",
-		"CMS167",
-		"CMS169",
-		"CMS177",
-		"CMS179",
-		"CMS182"
-	));
-	
-	protected static final List<String> AMBULATORY_CORE_CQMS = new ArrayList<String>(Arrays.asList(
-		"CMS2",
-		"CMS50",
-		"CMS68",
-		"CMS69",
-		"CMS75",
-		"CMS90",
-		"CMS117",
-		"CMS126",
-		"CMS136",
-		"CMS138",
-		"CMS146",
-		"CMS153",
-		"CMS154",
-		"CMS155",
-		"CMS156",
-		"CMS165",
-		"CMS166"	
-	));
-
 	public Validator20142015() {
-		this.counts.put("criteriaRequired", 0);
+		this.counts.put("criteriaRequired", 0);	// This number is calculated during the checks
 		this.counts.put("criteriaRequiredMet", 0);
-		this.counts.put("cqmsInpatientRequired", 16);
+		this.counts.put("cqmsInpatientRequired", 0);
 		this.counts.put("cqmsInpatientRequiredMet", 0);
-		this.counts.put("cqmsAmbulatoryRequired", 3);
+		this.counts.put("cqmsAmbulatoryRequired", 0);
 		this.counts.put("cqmsAmbulatoryRequiredMet", 0);
-		this.counts.put("cqmsAmbulatoryCoreRequired", 6);
+		this.counts.put("cqmsAmbulatoryCoreRequired", 0);
 		this.counts.put("cqmsAmbulatoryCoreRequiredMet", 0);
-		this.counts.put("domainsRequired", 3);
+		this.counts.put("domainsRequired", 0);
 		this.counts.put("domainsRequiredMet", 0);
 	}
 	
@@ -489,71 +370,18 @@ public class Validator20142015 extends Validator {
 	//**********************************************************************
 	// isCqmsValid
 	//
-	// Either Inpatient or Ambulatory CQMs required.
+	// There is no CQM check for 2014/2015 hybrid attestation.
 	//**********************************************************************
 	protected boolean isCqmsValid() {
-		boolean valid = false;
-		if (isAmbulatoryCqmsValid()) {
-			valid = true;
-		}
-		if (isInpatientCqmsValid()) {
-			valid = true;
-		}
-		return valid;
+		return true;
 	}
 
 	//**********************************************************************
 	// isDomainsValid
 	//
-	// At least 3 CQM Domains must be met.
+	// There is no CQM check for 2014/2015 hybrid attestation.
 	//**********************************************************************
 	protected boolean isDomainsValid() {
-		this.counts.put("domainsRequiredMet", 
-			this.domainsMet.size() >= this.counts.get("domainsRequired") ?
-				this.counts.get("domainsRequired") : this.domainsMet.size());
-		return (this.counts.get("domainsRequiredMet") >= this.counts.get("domainsRequired"));
-	}
-
-	//**********************************************************************
-	// isInpatientCqmsValid
-	//
-	// At least 16 Inpatient CQMs must be met.
-	//**********************************************************************
-	protected boolean isInpatientCqmsValid() {
-		int cqmCount = 0;
-		for (String cqm : INPATIENT_CQMS) {
-			if (this.cqmsMet.containsKey(cqm)) {
-				++cqmCount;
-			}
-		}
-		
-		this.counts.put("cqmsInpatientRequiredMet", cqmCount);
-		return (this.counts.get("cqmsInpatientRequiredMet") >= this.counts.get("cqmsInpatientRequired"));
-	}
-
-	//**********************************************************************
-	// isAmbulatoryCqmsValid
-	//
-	// At least 9 total Ambulatory CQMs with at least 6 of those being 
-	// Ambulatory Core CQMs.
-	//**********************************************************************
-	protected boolean isAmbulatoryCqmsValid() {
-		int nonCoreAmbulatory = 0;
-		int coreAmbulatory = 0;
-
-		for (String cqm : cqmsMet.keySet()) {
-			if (this.AMBULATORY_CORE_CQMS.contains(cqm)) {
-				++coreAmbulatory;
-			}
-			if (this.AMBULATORY_CQMS.contains(cqm)) {
-				++nonCoreAmbulatory;
-			}
-		}
-		
-		this.counts.put("cqmsAmbulatoryRequiredMet", nonCoreAmbulatory);
-		this.counts.put("cqmsAmbulatoryCoreRequiredMet", coreAmbulatory);
-		
-		return ((this.counts.get("cqmsAmbulatoryRequiredMet") + this.counts.get("cqmsAmbulatoryCoreRequiredMet")) >= 
-			(this.counts.get("cqmsAmbulatoryRequiredMet") + this.counts.get("cqmsAmbulatoryCoreRequiredMet")));
+		return true;
 	}
 }